fix: use layout as container
