--- client/src/components/Layout/Layout.tsx
@@ -1,12 +1,0 @@
-import Header from "../Header";
-
-const Layout: React.FC<{}> = ({ children }) => {
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-    </>
-  );
-};
-
-export default Layout;
