+++ client/src/pages/DashboardPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/Layout";
+import Layout from "../containers/Layout";
 
 const DashboardPage: React.FC<{}> = () => {
   return (
