+++ client/src/pages/LoadingPage.tsx
@@ -1,4 +1,5 @@
-import Layout from "../components/Layout";
+import React from "react";
+import Layout from "../containers/Layout";
 
 const LoadingPage: React.FC<{}> = () => {
   return (
